+++ RELEASE_NOTES.md
@@ -1,3 +1,11 @@
+## Version 0.1.1 - 5/24/2020
+ * Fix some console output
+ * Add diff so already present labels are not attempted to be created
+
+## Version 0.1.0 - 5/24/2020
+ * Added label display for repos
+ * Added label copy from one repo to another
+
 ## Version 0.0.1 - 5/24/2020
  * Test